Readjusted the menu, will test on secondary device then move on if it's proper

+++ MainMenu.cpp
@@ -3,28 +3,6 @@
 #include "MainMenu.h"
 #include "InputSystem.h"
 
-// static const float m_titleYSizeFraction = 0.2f;
-// static const sf::Vector2f m_titlePosFraction = {0.3f, 0.05f};
-    // static const float m_titleYPosFraction = 0.3f;
-    // static const float m_titleXPosFraction = 0.5f;
-//
-// static const float m_buttonYposFraction = 0.5f;
-// static const float m_buttonXposFraction = 0.5f;
-// static const float m_buttonYposMargin = 0.05f; // space between buttons vertically
-// static const float m_buttonYMargin = 0.02; // space between text and border of button
-// static const float m_buttonWidth = 0.35f;
-// static const float m_buttonHeight = 0.04f;
-//
-// static const sf::Color m_titleTextColor = sf::Color::Green;
-// static const sf::Color m_buttonTextColor = sf::Color::Green;
-// static const sf::Color m_buttonHvTextColor = sf::Color::White;
-// static const sf::Color m_buttonBgColor = sf::Color::White;
-// static const sf::Color m_buttonHvBgColor = sf::Color::Green;
-// static const sf::Color m_buttonOlColor = sf::Color::Green;
-
-// MenuButton(sf::Font font, std::string str, float marginX, float marginY, 
-//            sf::Color bgColor, sf::Color hvbgColor, sf::Color txColor, sf::Color hvtxColor, sf::Color olColor);
-
 MainMenu::MainMenu(const sf::Font& font)
 : m_title(font, "Momentum", m_titleTextColor, m_titlePosFraction, m_titleYSizeFraction), m_MenuButtons(), m_selection(0){
     m_MenuButtons.push_back(MenuButton(font, "Play", {m_buttonWidth, m_buttonHeight}, m_buttonXMargin, m_buttonYMargin,
@@ -65,8 +43,8 @@ void MainMenu::update(sf::Vector2i mousePos, InputSystem& inputSystem){
     }
 
     // update the selected button/deselect others
-    for(int i = 0; i < m_MenuButtons.size(); i++){
-        if(i == m_selection){
+    for(size_t i = 0; i < m_MenuButtons.size(); i++){
+        if(static_cast<int>(i) == m_selection){
             m_MenuButtons[i].select();
         } else m_MenuButtons[i].unselect();
     }
